CPR-838 Add more tests to e2e
